This PR adds a bash script to create a pgo optimized version of evmrs. In order to to so the benchmarks have to dump instumentation data.
Unfortunately this does not work if feature tail-call is enabled because in segfaults, likely because the tail calls are no longer optimized away when instrumented and therefore trigger a stack overflow.

```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Xeon(R) CPU E5-2690 v2 @ 3.00GHz
                        │ without-pgo  │              with-pgo               │
                        │    sec/op    │   sec/op     vs base                │
StaticOverhead/1/evmrs     2.490µ ± 1%   2.401µ ± 0%   -3.57% (p=0.000 n=20)
Inc/1/evmrs                5.835µ ± 3%   5.668µ ± 1%   -2.87% (p=0.000 n=20)
Inc/10/evmrs               6.011µ ± 2%   5.677µ ± 1%   -5.56% (p=0.000 n=20)
Fib/1/evmrs                5.429µ ± 1%   5.135µ ± 0%   -5.42% (p=0.000 n=20)
Fib/5/evmrs                23.65µ ± 1%   22.66µ ± 0%   -4.17% (p=0.000 n=20)
Fib/10/evmrs               244.4µ ± 0%   239.6µ ± 0%   -1.95% (p=0.000 n=20)
Fib/15/evmrs               2.402m ± 0%   2.346m ± 0%   -2.31% (p=0.000 n=20)
Fib/20/evmrs               26.41m ± 0%   25.54m ± 0%   -3.29% (p=0.000 n=20)
Sha3/1/evmrs               3.010µ ± 0%   2.770µ ± 0%   -7.99% (p=0.000 n=20)
Sha3/10/evmrs              4.928µ ± 1%   4.328µ ± 0%  -12.17% (p=0.000 n=20)
Sha3/100/evmrs             22.73µ ± 0%   19.69µ ± 0%  -13.34% (p=0.000 n=20)
Sha3/1000/evmrs            222.1µ ± 0%   193.6µ ± 0%  -12.82% (p=0.000 n=20)
Arith/1/evmrs              6.946µ ± 1%   6.409µ ± 1%   -7.73% (p=0.000 n=20)
Arith/10/evmrs             15.70µ ± 0%   14.51µ ± 0%   -7.60% (p=0.000 n=20)
Arith/100/evmrs            111.6µ ± 0%   105.4µ ± 0%   -5.55% (p=0.000 n=20)
Arith/280/evmrs            300.4µ ± 0%   279.1µ ± 0%   -7.09% (p=0.000 n=20)
Memory/1/evmrs             9.464µ ± 1%   8.885µ ± 1%   -6.12% (p=0.000 n=20)
Memory/10/evmrs            19.57µ ± 0%   17.89µ ± 0%   -8.61% (p=0.000 n=20)
Memory/100/evmrs           130.9µ ± 0%   120.4µ ± 0%   -7.98% (p=0.000 n=20)
Memory/1000/evmrs         1086.0µ ± 0%   998.6µ ± 0%   -8.05% (p=0.000 n=20)
Memory/10000/evmrs        10.463m ± 0%   9.621m ± 0%   -8.05% (p=0.000 n=20)
Analysis/jumpdest/evmrs    2.460µ ± 1%   2.369µ ± 0%   -3.72% (p=0.000 n=20)
Analysis/stop/evmrs        2.471µ ± 0%   2.372µ ± 0%   -4.03% (p=0.000 n=20)
Analysis/push1/evmrs       2.468µ ± 1%   2.367µ ± 0%   -4.09% (p=0.000 n=20)
Analysis/push32/evmrs      2.463µ ± 0%   2.369µ ± 1%   -3.82% (p=0.000 n=20)
geomean                    35.27µ        33.02µ        -6.37%
```